tune: refactor tuning to put ray in a submodule
