+++ src/lenskit/tuning/_ray/stopper.py
